+++ src/api/c/parser.cpp
@@ -101,7 +101,7 @@ bitwuzla_parser_get_bitwuzla(BitwuzlaParser* parser)
   BITWUZLA_CHECK_NOT_NULL(parser);
   if (!parser->d_bitwuzla)
   {
-    parser->d_bitwuzla.reset(new Bitwuzla(parser->d_parser->bitwuzla()));
+    parser->d_bitwuzla.reset(new Bitwuzla(parser->d_parser->bitwuzla().get()));
   }
   res = parser->d_bitwuzla.get();
   BITWUZLA_TRY_CATCH_END;